Centralize consts in files they can be imported from w/o downstream imports
