+++ src/games/cowpoke/cowpoke-game-object-types.ts
@@ -1,0 +1,13 @@
+export enum CHARACTER_TYPES {
+  UNASSIGNED = -1,
+  PLAYER = 0,
+  ENEMY = 1,
+}
+
+export const DECOR_TYPES = {
+  UNASSIGNED: -1,
+  BACK: 0,
+  MID: 1,
+  FLOOR: 2,
+  FRONT: 3,
+};
