+++ src/games/cowpoke/dialog.ts
@@ -1,0 +1,33 @@
+export const CHARACTER_KILL_QUIPS: string[] = [
+  "You've yee'd yer last haw, partner",
+  "This town ain't big enough for the tew of us",
+  "Looks like you brought a lasso to a gunfight",
+  "Guess you ran outta luck, cowboy",
+  "Shoulda stayed on the porch, partner",
+  "Hope you said all yer howdys and goodbyes",
+  "That's the last tumbleweed you'll chase",
+  "You bit the cactus, compadre",
+  "You done rode yer last rodeo",
+  "You couldn't dodge this bullet, friend",
+];
+
+export const PLAYER_DIALOG_QUIPS: string[] = [
+  "Looks like trouble's brewin' on the horizon",
+  "Someone's lookin' to make a name for themselves",
+  "Looks like there's a new hat in town",
+  "Reckon things just got a bit more interestin'",
+  "Better keep yer hand close to yer holster",
+];
+
+export const ENEMY_DIALOG_QUIPS: string[] = [
+  "Well would you look what the tumbleweed dragged in",
+  "You best be ready, partner. I'm fixin' to cause some trouble",
+  "Ain't no lawman gonna save you now",
+  "Time to see who's the fastest draw in these parts",
+  "I heard you were tough. Let's see if that's true",
+  "You lookin' for a showdown? 'Cause you just found one",
+  "Don't blink, or you'll miss yer last sunrise",
+  "Hope yer boots are fast, 'cause my bullets are faster",
+  "You look nervous, partner. That's a good instinct",
+  "Let's see if you can dance when the bullets start flyin'",
+];
